refactor: route관련 함수명 명확하게 수정
