+++ src/components/Header/Header.ts
@@ -1,5 +1,5 @@
 import Component from '../template/Component';
-import { push } from '../../routes/routeUtils';
+import { push } from '../../routes/router';
 export default class Header extends Component<undefined, undefined> {
   template() {
     return `<h1 data-name="home-header">Vanilla TS Projects</h1>`;
